Fix recall visibility, VecScore type, embedding dims; improve memories UI; add Codex skill
Knowledge recall:
- Drop s.kind constraint from visibility checks — no team-kind scopes exist,
  causing all team/department-visibility artifacts to silently return nothing;
  use path ancestry check alone for both visibility levels
- Fix VecScore field typed as int32 in RecallArtifactsByVectorRow,
  RecallMemoriesByVectorRow, RecallMemoriesByCodeVectorRow — should be float32

Embedding:
- Migration 000008: resize embedding_code from vector(1024) to vector(1536)
  to match text-embedding-3-small used for both text and code models;
  rebuilds HNSW index accordingly

Memories UI:
- Load all scopes as dropdown instead of raw UUID text input
- Default to first scope so memories load on initial page visit
- Wire offset-based pagination (page size 50) and populate NextCursor
  so the Load more button actually works

Graph:
- Remove LIMIT/OFFSET from ListRelationsByScope — graph needs all edges
- Style same_as edges distinctly (orange dashed, tighter spring)

Codex skill:
- Add .codex/skills/postbrain.md covering server reachability check,
  scope resolution, session lifecycle, and all 16 MCP tools with
  parameter tables and workflow guidance
- Add scripts/install-codex-skill.sh to copy skill into any project

+++ internal/db/memories.sql.go
@@ -392,7 +392,7 @@ type RecallMemoriesByCodeVectorRow struct {
 	SourceRef            *string
 	CreatedAt            time.Time
 	UpdatedAt            time.Time
-	VecScore             int32
+	VecScore             float32
 }
 
 func (q *Queries) RecallMemoriesByCodeVector(ctx context.Context, arg RecallMemoriesByCodeVectorParams) ([]*RecallMemoriesByCodeVectorRow, error) {
@@ -673,7 +673,7 @@ type RecallMemoriesByVectorRow struct {
 	SourceRef            *string
 	CreatedAt            time.Time
 	UpdatedAt            time.Time
-	VecScore             int32
+	VecScore             float32
 }
 
 func (q *Queries) RecallMemoriesByVector(ctx context.Context, arg RecallMemoriesByVectorParams) ([]*RecallMemoriesByVectorRow, error) {
